LIBASPACE-81 Remove calling hostname for every page request
This removes the need to call hostname for every request and DRYs out
the umd_environment_banner_helper

https://issues.umd.edu/browse/LIBASPACE-81
